fix: removing download of Roboto font
None of our other MFEs do this; it was left over from an early design decision.  It also doesn't appear to be having any effect, as I think this font is included via Paragon, if at all.  It may also be that it's a secondary font choice after something else already available.  Regardless, removing these two lines had no effect on the look and feel of the MFE, and reduces bundle size.