fix: modernize build process, remove npm publish, use github actions (#472)
* build: switch from travis to Github actions.

Subsequent commits will clean up some of the 'npm-dist' code that we're no longer using.

* build: remove npm distribution from repository

Once upon a time, we published the components of the profile MFE to NPM because we believed we might want to embed them in frontend-app-learner-portal.  We abandoned that idea about two years ago, but never cleaned up the experiment of publishing the MFE to npm.

This commit does that.  The prior commit on this branch disabled it by cutting over from travis to github actions - this commit just finishes the job.

* fix: cleaning up a bad URL for order history.

* docs: cleaning up a documentation URL in openedx.yml

* fix: removing download of Roboto font

None of our other MFEs do this; it was left over from an early design decision.  It also doesn't appear to be having any effect, as I think this font is included via Paragon, if at all.  It may also be that it's a secondary font choice after something else already available.  Regardless, removing these two lines had no effect on the look and feel of the MFE, and reduces bundle size.

* fix: using updated logo instead of pre-rebrand edX logo.

This now uses the FAVICON_URL environment variable like all our other MFEs.